+++ src/main/java/com/github/jlangch/aviron/ex/NotRunningException.java
@@ -1,0 +1,12 @@
+package com.github.jlangch.aviron.ex;
+
+
+public class NotRunningException extends AvironException {
+
+    public NotRunningException(final String message) {
+        super(message);
+    }
+
+
+    private static final long serialVersionUID = 1L;
+}