+++ src/test/java/com/github/jlangch/aviron/TestManagementFunctions.java
@@ -45,7 +45,7 @@ public class TestManagementFunctions {
         System.out.println();
         
         System.out.println("[clamd PID]");
-        System.out.println(client.getClamdPID());
+        System.out.println(Admin.getClamdPID());
         System.out.println();
         System.out.println();
     }